dalle-nft-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
web3==6.10.0
openai==1.3.0
requests==2.31.0
